+++ fortitude/src/configuration.rs
@@ -1,6 +1,8 @@
 use crate::cli::CheckArgs;
 use crate::fs::{FilePattern, FilePatternSet, EXCLUDE_BUILTINS, FORTRAN_EXTS};
-use crate::options::{ExitUnlabelledLoopOptions, Options};
+use crate::options::{
+    ExitUnlabelledLoopOptions, KeywordWhitespaceOptions, Options, PortabilityOptions, StringOptions,
+};
 use crate::registry::RuleNamespace;
 use crate::rule_redirects::get_redirect;
 use crate::rule_selector::{
@@ -167,6 +169,9 @@ pub struct Configuration {
     pub gitignore_mode: GitignoreMode,
     // Individual rules
     pub exit_unlabelled_loops: Option<ExitUnlabelledLoopOptions>,
+    pub keyword_whitespace: Option<KeywordWhitespaceOptions>,
+    pub strings: Option<StringOptions>,
+    pub portability: Option<PortabilityOptions>,
 }
 
 impl Default for Configuration {
@@ -191,6 +196,9 @@ impl Default for Configuration {
             exclude_mode: Default::default(),
             gitignore_mode: Default::default(),
             exit_unlabelled_loops: Default::default(),
+            keyword_whitespace: Default::default(),
+            strings: Default::default(),
+            portability: Default::default(),
         }
     }
 }
@@ -261,6 +269,9 @@ impl Configuration {
 
             // Individual rules
             exit_unlabelled_loops: check.exit_unlabelled_loops,
+            keyword_whitespace: check.keyword_whitespace,
+            strings: check.strings,
+            portability: check.portability,
         }
     }
 
@@ -359,6 +370,18 @@ impl Configuration {
                 exit_unlabelled_loops: self
                     .exit_unlabelled_loops
                     .map(ExitUnlabelledLoopOptions::into_settings)
+                    .unwrap_or_default(),
+                keyword_whitespace: self
+                    .keyword_whitespace
+                    .map(KeywordWhitespaceOptions::into_settings)
+                    .unwrap_or_default(),
+                strings: self
+                    .strings
+                    .map(StringOptions::into_settings)
+                    .unwrap_or_default(),
+                portability: self
+                    .portability
+                    .map(PortabilityOptions::into_settings)
                     .unwrap_or_default(),
             },
             file_resolver: FileResolverSettings {