+++ fortitude/src/rules/correctness/mod.rs
@@ -1,5 +1,6 @@
 pub mod accessibility_statements;
 pub mod assumed_size;
+pub mod conditionals;
 pub mod derived_default_init;
 pub mod exit_labels;
 pub mod external;
@@ -10,17 +11,23 @@ pub mod intent;
 pub mod kind_suffixes;
 pub mod magic_numbers;
 pub mod missing_io_specifier;
+pub mod nonportable_shortcircuit_inquiry;
 pub mod select_default;
+pub mod split_escaped_quote;
 pub mod trailing_backslash;
 pub mod use_statements;
 
 #[cfg(test)]
 mod tests {
     use std::convert::AsRef;
+    use std::fs;
     use std::path::Path;
+    use std::process::Command;
 
     use anyhow::Result;
+    use assert_cmd::prelude::*;
     use insta::assert_snapshot;
+    use tempfile::TempDir;
     use test_case::test_case;
 
     use crate::apply_common_filters;
@@ -43,6 +50,7 @@ mod tests {
     #[test_case(Rule::AssumedSize, Path::new("C071.f90"))]
     #[test_case(Rule::AssumedSizeCharacterIntent, Path::new("C072.f90"))]
     #[test_case(Rule::InitialisationInDeclaration, Path::new("C081.f90"))]
+    #[test_case(Rule::PointerInitialisationInDeclaration, Path::new("C082.f90"))]
     #[test_case(Rule::ExternalProcedure, Path::new("C091.f90"))]
     #[test_case(Rule::ProcedureNotInModule, Path::new("C092.f90"))]
     #[test_case(Rule::MissingDefaultPointerInitalisation, Path::new("C101.f90"))]
@@ -52,6 +60,11 @@ mod tests {
     #[test_case(Rule::DefaultPublicAccessibility, Path::new("C132.f90"))]
     #[test_case(Rule::MissingExitOrCycleLabel, Path::new("C141.f90"))]
     #[test_case(Rule::ExitOrCycleInUnlabelledLoop, Path::new("C142.f90"))]
+    #[test_case(Rule::MissingEndLabel, Path::new("C143.f90"))]
+    #[test_case(Rule::MisleadingInlineIfSemicolon, Path::new("C151.f90"))]
+    #[test_case(Rule::MisleadingInlineIfContinuation, Path::new("C152.f90"))]
+    #[test_case(Rule::NonportableShortcircuitInquiry, Path::new("C161.f90"))]
+    #[test_case(Rule::SplitEscapedQuote, Path::new("C171.f90"))]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
         let snapshot = format!("{}_{}", rule_code.as_ref(), path.to_string_lossy());
         let diagnostics = test_path(
@@ -106,6 +119,32 @@ mod tests {
         )?;
         apply_common_filters!();
         assert_snapshot!(snapshot, diagnostics);
+        Ok(())
+    }
+
+    #[test]
+    fn c151_fix_multiple_inline_if() -> Result<()> {
+        let tempdir = TempDir::new()?;
+        let path = tempdir.path().join("C151.f90");
+        let code = r#"
+        program test
+            implicit none
+            integer :: i
+            if (i == 1) print *, "foo"; if (i == 2) print *, "bar"; if (i == 3) print *, "baz"
+        end program test
+        "#;
+        fs::write(&path, code)?;
+        Command::cargo_bin("fortitude")?
+            .arg("check")
+            .arg("--fix")
+            .arg("--preview")
+            .arg("--select=C151")
+            .arg(path.as_os_str())
+            .status()?;
+        let fixed = String::from_utf8(fs::read(path.as_os_str())?)?;
+        let snapshot = "c151_fix_multiple_inline_if";
+        apply_common_filters!();
+        assert_snapshot!(snapshot, fixed);
         Ok(())
     }
 }