+++ fortitude/src/rules/modernisation/include_statement.rs
@@ -1,9 +1,9 @@
-use crate::settings::Settings;
 use crate::AstRule;
+use crate::{ast::FortitudeNode, settings::Settings};
 use ruff_diagnostics::{Diagnostic, Violation};
 use ruff_macros::{derive_message_formats, ViolationMetadata};
 use ruff_source_file::SourceFile;
-use ruff_text_size::{TextRange, TextSize};
+use ruff_text_size::TextRange;
 use tree_sitter::Node;
 
 /// ## What it does
@@ -36,8 +36,8 @@ impl AstRule for IncludeStatement {
     fn check(_settings: &Settings, node: &Node, _src: &SourceFile) -> Option<Vec<Diagnostic>> {
         // tree-sitter-fortran 0.5.1 includes the end newline as part
         // of the node, so we discard that here
-        let start = TextSize::try_from(node.child(0)?.start_byte()).unwrap();
-        let end = TextSize::try_from(node.child(1)?.end_byte()).unwrap();
+        let start = node.child(0)?.start_textsize();
+        let end = node.child(1)?.end_textsize();
         let range = TextRange::new(start, end);
         some_vec![Diagnostic::new(IncludeStatement {}, range)]
     }