+++ fortitude/src/rules/style/double_colon_in_decl.rs
@@ -4,7 +4,6 @@ use crate::{AstRule, FromAstNode};
 use ruff_diagnostics::{AlwaysFixableViolation, Diagnostic, Edit, Fix};
 use ruff_macros::{derive_message_formats, ViolationMetadata};
 use ruff_source_file::SourceFile;
-use ruff_text_size::TextSize;
 use tree_sitter::Node;
 
 /// ## What does it do?
@@ -34,7 +33,7 @@ impl AstRule for MissingDoubleColon {
             .all(|child| child != "::")
         {
             let first_decl = node.child_by_field_name("declarator")?;
-            let start_pos = TextSize::try_from(first_decl.start_byte()).unwrap();
+            let start_pos = first_decl.start_textsize();
             let fix = Fix::safe_edit(Edit::insertion(":: ".to_string(), start_pos));
             some_vec!(Diagnostic::from_node(Self {}, node).with_fix(fix))
         } else {