+++ fortitude/src/settings.rs
@@ -20,6 +20,8 @@ use crate::registry::Rule;
 use crate::rule_selector::{CompiledPerFileIgnoreList, PreviewOptions, RuleSelector};
 use crate::rule_table::RuleTable;
 use crate::rules::correctness::exit_labels;
+use crate::rules::portability::{self};
+use crate::rules::style::{keywords, strings};
 
 #[derive(Debug)]
 pub struct Settings {
@@ -71,6 +73,9 @@ pub struct CheckSettings {
 
     // Individual rule settings
     pub exit_unlabelled_loops: exit_labels::settings::Settings,
+    pub keyword_whitespace: keywords::settings::Settings,
+    pub strings: strings::settings::Settings,
+    pub portability: portability::settings::Settings,
 }
 
 impl CheckSettings {
@@ -92,6 +97,9 @@ impl CheckSettings {
             preview: PreviewMode::default(),
             ignore_allow_comments: IgnoreAllowComments::default(),
             exit_unlabelled_loops: exit_labels::settings::Settings::default(),
+            keyword_whitespace: keywords::settings::Settings::default(),
+            strings: strings::settings::Settings::default(),
+            portability: portability::settings::Settings::default(),
         }
     }
 }
@@ -121,6 +129,9 @@ impl fmt::Display for CheckSettings {
             namespace = "check",
             fields = [
                 self.exit_unlabelled_loops | nested,
+                self.keyword_whitespace | nested,
+                self.strings | nested,
+                self.portability | nested,
             ]
         }
         Ok(())